+++ dmmarkinserter.php
@@ -53,14 +53,14 @@ if (isset($data['studentName'])) {
             class VARCHAR(255) NOT NULL,
             term VARCHAR(100) NOT NULL,
           
-            test1 INT DEFAULT 0,
-            test2 INT DEFAULT 0,
-            Assignment1 INT DEFAULT 0,
-            Assignment2 INT DEFAULT 0,
-            GroupAssignment INT DEFAULT 0,
+            test1 DECIMAL(10,1) DEFAULT 0,
+            test2 DECIMAL(10,1) DEFAULT 0,
+            Assignment1 DECIMAL(10,1) DEFAULT 0,
+            Assignment2 DECIMAL(10,1) DEFAULT 0,
+            GroupAssignment DECIMAL(10,1) DEFAULT 0,
             
-            finalexam INT DEFAULT 0,
-            total INT DEFAULT 0,
+            finalexam DECIMAL(10,1) DEFAULT 0,
+            total DECIMAL(10,1) DEFAULT 0,
             date DATE NOT NULL,
             UNIQUE(student_name, term)
         )";
